Refactored CommitteeController
