+++ app/Resources/views/admin/committee_actions.html.twig
@@ -18,7 +18,7 @@
     </div>
 
     <div style="width: 160px; margin-top: 5px;">
-        <a href="{{ path('admin_app_committee_edit', {id: object.id}) }}" class="btn btn-xs btn-default" title="Éditer">
+        <a href="{{ path('app_committee_manager_edit', {id: object.id}) }}" class="btn btn-xs btn-default" title="Éditer">
             <i class="fa fa-pencil" aria-hidden="true"></i>
             Éditer
         </a>
