+++ app/Resources/views/committee/_host_header.html.twig
@@ -3,13 +3,13 @@
 <div id="committee-host-nav" class="committee__host__nav">
     <ul class="list__links list__links--default list__links--row list__links--row-left">
         <li>
-            <a href="{{ committee_path('app_committee_edit', committee) }}" class="icon--with-text">
+            <a href="{{ committee_path('app_committee_manager_edit', committee) }}" class="icon--with-text">
                 <i class="fa fa-sliders"></i>
                 Éditer le comité
             </a>
         </li>
         <li>
-            <a href="{{ committee_path('app_commitee_list_members', committee) }}" class="icon--with-text">
+            <a href="{{ committee_path('app_commitee_manager_list_members', committee) }}" class="icon--with-text">
                 <i class="fa fa-users"></i>
                 Gérer les adhérents
             </a>
@@ -23,7 +23,7 @@
             Message à tous les adhérents du comité
         </li>
         <li>
-            <a href="{{ committee_path('app_committee_add_event', committee) }}" class="link--no-decor">
+            <a href="{{ committee_path('app_committee_manager_add_event', committee) }}" class="link--no-decor">
                 Créer un événement
             </a>
         </li>
