+++ app/Resources/views/committee/contact.html.twig
@@ -17,7 +17,7 @@
 {% block committee_content %}
     <main>
         <section class="l__wrapper--medium space--30-0">
-            <a href="{{ committee_path('app_commitee_list_members', committee) }}" class="back-to-list icon--with-text">
+            <a href="{{ committee_path('app_commitee_manager_list_members', committee) }}" class="back-to-list icon--with-text">
                 {{ include('components/caret--left.html.twig') }}
                 Retour à la liste</a>
 
@@ -48,7 +48,7 @@
 
                 {{ form_widget(form.submit, { label: 'Envoyer', attr: { class: 'btn btn--blue b__nudge--top-15' } }) }}
 
-                <a href="{{ committee_path('app_commitee_list_members', committee) }}" class="btn">
+                <a href="{{ committee_path('app_commitee_manager_list_members', committee) }}" class="btn">
                     Annuler
                 </a>
             {{ form_end(form) }}
