+++ app/Resources/views/committee/demote_host.html.twig
@@ -24,7 +24,7 @@
 
         <div class="l__wrapper l__row l__row--top l__row--h-stretch l__tablet--col">
             <div class="committee__main l__wrapper--main l__wrapper--main-left">
-                <a href="{{ committee_path('app_commitee_list_members', committee) }}"
+                <a href="{{ committee_path('app_commitee_manager_list_members', committee) }}"
                    class="back-to-list icon--with-text text--summary space--20-0">
                     {{ include('components/caret--left.html.twig') }}
                     Retour aux membres du comité
@@ -48,7 +48,7 @@
                             <button type="submit" id="members-promote-button" class="btn btn--primary b__nudge--right">
                                 Oui, définir comme simple membre
                             </button>
-                            <a href="{{ committee_path('app_commitee_list_members', committee) }}" class="btn">
+                            <a href="{{ committee_path('app_commitee_manager_list_members', committee) }}" class="btn">
                                 Non, annuler
                             </a>
                         </form>
