+++ app/Resources/views/committee/hosts.html.twig
@@ -21,5 +21,5 @@
 
 <p class="text--summary">
     Pour ajouter ou supprimer des animateurs, dirigez-vous vers la
-    <a href="{{ path('app_commitee_list_members', {'uuid': committee.uuid.toString, 'slug': committee.slug}) }}">section des membres</a>.
+    <a href="{{ path('app_commitee_manager_list_members', {'uuid': committee.uuid.toString, 'slug': committee.slug}) }}">section des membres</a>.
 </p>
