+++ app/Resources/views/committee/show.html.twig
@@ -179,7 +179,7 @@
                 <ul class="list__links list__links--default list__links--row list__links--row-left"></ul>
                 <ul class="list__links list__links--default list__links--row">
                     <li>
-                        <a href="{{ committee_path('app_committee_edit', committee) }}" class="icon--with-text">
+                        <a href="{{ committee_path('app_committee_manager_edit', committee) }}" class="icon--with-text">
                             <i class="fa fa-sliders"></i>
                             Éditer le comité
                         </a>
