+++ app/Resources/views/committee_manager/list_members.html.twig
@@ -132,7 +132,7 @@
                             <button id="members-contact-button" class="btn b__nudge--right" disabled="disabled">Contacter</button>
                         </form>
 
-                        <form method="post" action="{{ committee_path('app_commitee_export_members', committee) }}">
+                        <form method="post" action="{{ committee_path('app_commitee_manager_export_members', committee) }}">
                             <input type="hidden" name="token" value="{{ csrf_token('committee.export_members') }}" id="members-export-token" />
                             <input type="hidden" name="exports" value="" id="members-export-selection" />
